API

Datasets for Clima Look-Up-Table

GriddingMachine.AbstractDataset — Type

abstract type AbstractDataset{FT}

Hierachy of AbstractDataset

MeanMonthlyLAI

GriddingMachine.MeanMonthlyLAI — Type

struct MeanMonthlyLAI{FT<:AbstractFloat}

A struct that contains monthly mean LAI

Fields

LAI

Monthly mean LAI

res_lat

Latitude resolution [°]

res_lon

Longitude resolution [°]

Function to read Look-Up-Table

GriddingMachine.lat_ind — Function

lat_ind(lat::FT; res::FT) where {FT<:AbstractFloat}

Round the latitude and return the index in a matrix, Given

lat Latitude

res Resolution in latitude

GriddingMachine.lon_ind — Function

lon_ind(lon::FT; res::FT=1) where {FT<:AbstractFloat}

Round the longitude and return the index in a matrix, Given

lon Longitude

res Resolution in longitude

GriddingMachine.read_LUT — Function

read_LUT(
            dataset::MeanMonthlyLAI{FT},
            lat::FT,
            lon::FT,
            month::Int
) where {FT<:AbstractFloat}

Read the LAI from given

dataset MeanMonthlyLAI type struct

lat Latitude

lon Longitude

month Month number from 1 to 12, or String