API

Datasets for CliMA Look-Up-Table

All data are stored in a general structure, but they are catergorized to different data types. The general data structure is

GriddingMachine.GriddedDataset — Type

struct GriddedDataset{FT<:AbstractFloat}

A general struct to store data

Fields

data

Gridded dataset

res_lat

Latitude resolution [°]

res_lon

Longitude resolution [°]

res_time

Time resolution: D-M-Y-C: day-month-year-century

var_name

Variable name

var_attr

Variable attribute

dt

Type label

Note it here that the dt field in GriddedDataset is the data identity of the stored data. Please refer the lists below

Dataset types

Current supported dataset formats are

GriddingMachine.AbstractFormat — Type

abstract type AbstractFormat

Hierachy of AbstractFormat

FormatTIFF

FormatNC

GriddingMachine.FormatNC — Type

struct FormatNC

GriddingMachine.FormatTIFF — Type

struct FormatTIFF

General type

GriddingMachine.AbstractDataset — Type

abstract type AbstractDataset{FT}

Hierachy of AbstractDataset

AbstractLeafDataset

AbstractStandDataset

Leaf level dataset

GriddingMachine.AbstractLeafDataset — Type

abstract type AbstractLeafDataset{FT}

Hierachy of AbstractLeafDataset

[LeafChlorophyll]9@ref

LeafNitrogen

LeafPhosphorus

LeafSLA

VcmaxOptimalCiCa

GriddingMachine.LeafChlorophyll — Type

struct LeafChlorophyll{FT}

Struct for leaf chlorophyll content

GriddingMachine.LeafNitrogen — Type

struct LeafNitrogen{FT}

Struct for leaf nitrogen content

GriddingMachine.LeafPhosphorus — Type

struct LeafPhosphorus{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area)

GriddingMachine.LeafSLA — Type

struct LeafSLA{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area)

GriddingMachine.VcmaxOptimalCiCa — Type

struct VcmaxOptimalCiCa{FT}

Struct for Vcmax estimated from optimal Ci:Ca ratio

Stand level datasets

GriddingMachine.AbstractStandDataset — Type

abstract type AbstractStandDataset{FT}

Hierachy of AbstractStandDataset

CanopyHeightGLAS

ClumpingIndexMODIS

ClumpingIndexPFT

GPPMPIv006

GPPVPMv20

LAIMonthlyMean

NPPModis

TreeDensity

GriddingMachine.CanopyHeightGLAS — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat

GriddingMachine.ClumpingIndexMODIS — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat

GriddingMachine.ClumpingIndexPFT — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat, for different plant functional types. The indices are Broadleaf, Needleleaf, C3 grasses, C4 grasses, and shrubland

GriddingMachine.GPPMPIv006 — Type

struct GPPMPIv006{FT}

Struct for MPI GPP v006

GriddingMachine.GPPVPMv20 — Type

struct GPPVPMv20{FT}

Struct for VPM GPP v20

GriddingMachine.LAIMonthlyMean — Type

struct LAIMonthlyMean{FT}

Struct for monthly mean MODIS LAI

GriddingMachine.LandMaskERA5 — Type

struct LandMaskERA5{FT}

Struct for land mask from ERA5

GriddingMachine.NPPModis — Type

struct NPPModis{FT}

Struct for Modis NPP

GriddingMachine.TreeDensity — Type

struct TreeDensity{FT}

Struct for tree density (number of trees per km⁻²)

LOAD and READ Look-Up-Table

Griddingmachine package allows for reading data from both the artifact and local files. To read the artifacts, function query_LUT is provided:

GriddingMachine.query_LUT — Function

query_LUT(dt::AbstractDataset)
query_LUT(dt::AbstractDataset, res_g::String)
query_LUT(dt::AbstractDataset, year::Int, res_g::String, res_t::String)

Query the file location from artifacts, given

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

Return the following:

Path to artifact data file

Format of the data

Variable name of NC file or Band of Tiff file

Time resolution

Whether latitude is reversed

Variable name of gridded data

Variable attribute of gridded data

A general function is provided to load the look-up tables. Before you load the table, you need to know what data you want to read (see AbstractDataset), data from which year (required for some datasets), geophysical resolution (required for some datasets), and time resolution (resuired for some datasets). Also, if you want to load local files, you need to know the file name, dataset format, and the dataset label (e.g., variable name in *.nc files, or band number in *.tif files). Be aware that some datasets may be very big, be patient when downloading the data files. Also, you need to use computers or servers with enough memory, otherwise the program may crash when loading the data.

GriddingMachine.load_LUT — Function

load_LUT(dt::AbstractDataset{FT}) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         res_g::String,
         res_t::String) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         year::Int,
         res_g::String,
         res_t::String) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         file::String,
         format::AbstractFormat,
         label::String,
         res_t::String,
         rev_lat::Bool,
         var_name::String,
         var_attr::Dict{String,String}) where {FT<:AbstractFloat}

Load look up table and return the struct, given

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

file File name to read, useful to read local files

format Dataset format from AbstractFormat

label Variable label in dataset, e.g., var name in .nc files, band numer in .tif files

rev_lat Whether latitude is stored reversely in the dataset, e.g., 90 to -90. If true, mirror the dataset on latitudinal direction

var_name Variable name of GriddedDataset

var_attr Variable attributes of GriddedDataset

Note that the artifact for GPP is about

500 MB for 0.2 degree resolution (5^2 * 36018046)

2600 MB for 0.083 degree resolution (12^2 * 36018046)

The loaded data may contain null values filled with different default values. To remove these unexpected data, a general function is provided to mask out these values.

GriddingMachine.mask_LUT! — Function

mask_LUT!(ds::GriddedDataset, default::Number)
mask_LUT!(ds::GriddedDataset, lims::Array)

Filter out the unrealistic values from the dataset, given

ds GriddedDataset type struct

default Optional. Default value to replace NaNs

lims Lower and upper limits for the dataset

A general function is provided to read the look-up tables. Note that you need to use realistic latitude, longitude, and index (required for some datasets) to read the data properly. For example, latitude must be within the range of [-90,90], longitude must be within the range of [-180,180] or [0,360], and index must be in [1,12] if you use it as month or [0,46] if you read the 8-day products. The functions lat_ind and lon_ind help convert latitude and longitude into index, respectively.

GriddingMachine.read_LUT — Function

read_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}

Read the dataset from given

ds GriddedDataset type struct

dt Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType

GriddingMachine.lat_ind — Function

lat_ind(lat::FT; res::FT) where {FT<:AbstractFloat}

Round the latitude and return the index in a matrix, Given

lat Latitude

res Resolution in latitude

GriddingMachine.lon_ind — Function

lon_ind(lon::FT; res::FT=1) where {FT<:AbstractFloat}

Round the longitude and return the index in a matrix, Given

lon Longitude

res Resolution in longitude

An alternative view method is provided in parallel with read_LUT. The difference is that there is no memory allocation using view_LUT so that changes can be made directly to the view, whereas read_LUT creates new number or arrays and changes cannot be made to original data directly.

GriddingMachine.view_LUT — Function

view_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}

View the dataset from given

ds GriddedDataset type struct

dt Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType

Regrid the Look-Up-Table

For better matching the gridded data, we also provide a function to regrid the dataset.

GriddingMachine.regrid_LUT — Function

regrid_LUT(ds::AbstractDataset{FT},
           zoom::Int;
           nan_weight::Bool=false
) where {FT<:AbstractFloat}

Regrid the data from high to low resolution and return the struct, given

ds GriddedDataset type struct

zoom Integer geophysical zoom factor (>=2)

nan_weight Optional. If true, assuming nan = 0; if false, neglecting it.

Save the Look-Up-Table

To avoid unnecessary reading or regridding the dataset, a function is provided to save the dataset to .nc file

GriddingMachine.save_LUT — Function

save_LUT(ds::GriddedDataset{FT}, filename::String) where {FT<:AbstractFloat}

Save the dataset, given

ds GriddedDataset type struct

filename NC file name to save to