API

Datasets for Clima Look-Up-Table

GriddingMachine.AbstractDataset — Type

abstract type AbstractDataset{FT}

Hierachy of AbstractDataset

AbstractGPP

AbstractLAI

GriddingMachine.LAIMonthlyMean — Type

struct LAIMonthlyMean{FT}

Struct for monthly mean MODIS LAI

Function to read Look-Up-Table

GriddingMachine.lat_ind — Function

lat_ind(lat::FT; res::FT) where {FT<:AbstractFloat}

Round the latitude and return the index in a matrix, Given

lat Latitude

res Resolution in latitude

GriddingMachine.lon_ind — Function

lon_ind(lon::FT; res::FT=1) where {FT<:AbstractFloat}

Round the longitude and return the index in a matrix, Given

lon Longitude

res Resolution in longitude

GriddingMachine.read_LUT — Function

read_LUT(data::GriddedDataset{FT}, lat::FT, lon::FT) where {FT<:AbstractFloat}
read_LUT(data::GriddedDataset{FT}, lat::FT, lon::FT, ind::Int) where {FT<:AbstractFloat}
read_LUT(data::GriddedDataset{FT}, data_type::AbstractGPP{FT}, lat::FT, lon::FT, ind::Int) where {FT<:AbstractFloat}
read_LUT(data::GriddedDataset{FT}, data_type::LAIMonthlyMean{FT}, lat::FT, lon::FT, ind::Int) where {FT<:AbstractFloat}
read_LUT(data::GriddedDataset{FT}, data_type::VcmaxOptimalCiCa{FT}, lat::FT, lon::FT) where {FT<:AbstractFloat}

Read the LAI from given

data LAIMonthlyMean type struct

data_type Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType