API

Collections

Gridded datasets

GriddingMachine.Collections.GriddedCollection — Type

struct GriddedCollection

Structure for general gridded dataset collection.

Fields

LABEL::String

Artifact label name

SUPPORTED_COMBOS::Vector{String}

Supported combinations

DEFAULT_COMBO::String

Default combination

Examples

vcmax_collection = GriddedCollection("VCMAX", ["2X_1Y_V1", "2X_1Y_V2"], "2X_1Y_V2");

GriddingMachine.Collections.canopy_height_collection — Function

canopy_height_collection()

Method to create a general dataset collection for canopy height. Supported datasets are (click to view bibtex items)

20X_1Y_V1 (Simard et al., 2011)

2X_1Y_V2 (Boonman et al., 2020)

@article{simard2011mapping,
    author = {Simard, Marc and Pinto, Naiara and Fisher, Joshua B and Baccini, Alessandro},
    year = {2011},
    title = {Mapping forest canopy height globally with spaceborne lidar},
    journal = {Journal of Geophysical Research: Biogeosciences},
    volume = {116},
    number = {G4021}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collections.clumping_index_collection — Function

clumping_index_collection()

Method to create a general dataset collection for clumping index. Supported datasets are (click to view bibtex items)

240X_1Y_V1 (He et al., 2012)

2X_1Y_V1 (regridded; He et al., 2012)

2X_1Y_V2 (Braghiere et al., 2019)

V2 dataset are classified for different plant functional types. The indices are Broadleaf, Needleleaf, C3 grasses, C4 grasses, and shrubland.

@article{he2012global,
    author={He, Liming and Chen, Jing M and Pisek, Jan and Schaaf, Crystal B and Strahler, Alan H},
    year={2012},
    title={Global clumping index map derived from the MODIS BRDF product},
    journal={Remote Sensing of Environment},
    volume={119},
    pages={118--130}
}
@article{braghiere2019underestimation,
    author = {Braghiere, R{\'e}nato Kerches and Quaife, T and Black, E and He, L and Chen, JM},
    year = {2019},
    title = {Underestimation of global photosynthesis in Earth System Models due to representation of vegetation structure},
    journal = {Global Biogeochemical Cycles},
    volume = {33},
    number = {11},
    pages = {1358--1369}
}

GriddingMachine.Collections.elevation_collection — Function

elevation_collection()

Method to create a general dataset collection for surface elevation. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Yamazaki et al., 2017)

@article{yamazaki2017high,
    author = {Yamazaki, Dai and Ikeshima, Daiki and Tawatari, Ryunosuke and Yamaguchi, Tomohiro and O'Loughlin, Fiachra and Neal, Jeffery C and Sampson, Christopher C and Kanae, Shinjiro and
              Bates, Paul D},
    year = {2017},
    title = {A high-accuracy map of global terrain elevations},
    journal = {Geophysical Research Letters},
    volume = {44},
    number = {11},
    pages = {5844--5853}
}

GriddingMachine.Collections.gpp_collection — Function

gpp_collection()

Method to create a general dataset collection for gross primary productivity. Supported datasets are (click to view bibtex items)

MPI_RS_2X_1M_YYYY_V1 (YYYY from 2001 to 2019; Tramontana et al., 2016)

MPI_RS_2X_8D_YYYY_V1 (YYYY from 2001 to 2019; Tramontana et al., 2016)

VPM_5X_8D_YYYY_V2 (YYYY from 2000 to 2019; Zhang et al., 2017)

VPM_12X_8D_YYYY_V2 (YYYY from 2000 to 2019; Zhang et al., 2017)

@article{tramontana2016predicting,
    author = {Tramontana, Gianluca and Jung, Martin and Schwalm, Christopher R and Ichii, Kazuhito and Camps-Valls, Gustau and R{'a}duly, Botond and Reichstein, Markus and Arain, M Altaf and
              Cescatti, Alessandro and Kiely, Gerard and others},
    year = {2016},
    title = {Predicting carbon dioxide and energy fluxes across global FLUXNET sites with regression algorithms},
    journal = {Biogeosciences},
    volume = {13},
    number = {14},
    pages = {4291--4313}
}
@article{zhang2017global,
    author = {Zhang, Yao and Xiao, Xiangming and Wu, Xiaocui and Zhou, Sha and Zhang, Geli and Qin, Yuanwei and Dong, Jinwei},
    year = {2017},
    title = {A global moderate resolution dataset of gross primary production of vegetation for 2000--2016},
    journal = {Scientific data},
    volume = {4},
    pages = {170165}
}

GriddingMachine.Collections.lai_collection — Function

lai_collection()

Method to create a general dataset collection for leaf area index. Supported datasets are (click to view bibtex items)

MODIS_2X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_2X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_10X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_10X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_20X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_20X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

@article{yuan2011reprocessing,
	author = {Yuan, Hua and Dai, Yongjiu and Xiao, Zhiqiang and Ji, Duoying and Shangguan, Wei},
	year = {2011},
	title = {Reprocessing the MODIS Leaf Area Index products for land surface and climate modelling},
	journal = {Remote Sensing of Environment},
	volume = {115},
	number = {5},
	pages = {1171--1187}
}

GriddingMachine.Collections.land_mask_collection — Function

land_mask_collection()

Method to create a general dataset collection for land mask. Supported datasets are (click to view bibtex items)

4X_1Y_V1 [(ERA5)]

GriddingMachine.Collections.leaf_chlorophyll_collection — Function

leaf_chlorophyll_collection()

Method to create a general dataset collection for leaf chlorophyll content. Supported datasets are (click to view bibtex items)

2X_7D_V1 (Croft et al., 2017)

@article{croft2020global,
    author = {Croft, H and Chen, JM and Wang, R and Mo, G and Luo, S and Luo, X and He, L and Gonsamo, A and Arabian, J and Zhang, Y and others},
    year = {2020},
    title = {The global distribution of leaf chlorophyll content},
    journal = {Remote Sensing of Environment},
    volume = {236},
    pages = {111479},
}

GriddingMachine.Collections.leaf_nitrogen_collection — Function

leaf_nitrogen_collection()

Method to create a general dataset collection for leaf nitrogen content. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

2X_1Y_V2 (Boonman et al., 2020)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collections.leaf_phosphorus_collection — Function

leaf_phosphorus_collection()

Method to create a general dataset collection for leaf phosphorus content. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}

GriddingMachine.Collections.pft_collection — Function

pft_collection()

Method to create a general dataset collection for plant function type ratio. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

@article{lawrence2007representing,
	author = {Lawrence, Peter J and Chase, Thomas N},
	year = {2007},
	title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
	journal = {Journal of Geophysical Research: Biogeosciences},
	volume = {112},
	pages = {G01023}
}

GriddingMachine.Collections.sif_collection — Function

sif_collection()

Method to create a general dataset collection for solar-induced chlorophyll fluorescence. Supported datasets are (click to view bibtex items)

TROPOMI_740_1X_1M_YYYY_V1 (YYYY from 2018 to 2019; Köhler et al., 2018)

TROPOMI_740_12X_8D_YYYY_V1 (YYYY from 2018 to 2019; Köhler et al., 2018)

TROPOMI_740DC_1X_1M_YYYY_V1 (YYYY from 2018 to 2019; Köhler et al., 2018)

TROPOMI_740DC_12X_8D_YYYY_V1 (YYYY from 2018 to 2019; Köhler et al., 2018)

@article{kohler2018global,
    author = {K{\"o}hler, Philipp and Frankenberg, Christian and Magney, Troy S and Guanter, Luis and Joiner, Joanna and Landgraf, Jochen},
    year = {2018},
    title = {Global retrievals of solar-induced chlorophyll fluorescence with {TROPOMI}: {F}irst results and intersensor comparison to {OCO-2}},
    journal = {Geophysical Research Letters},
    volume = {45},
    number = {19},
    pages = {10,456--10,463}
}

GriddingMachine.Collections.sla_collection — Function

sla_collection()

Method to create a general dataset collection for SLA (specific leaf area). Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

2X_1Y_V2 (Boonman et al., 2020)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collections.soil_color_collection — Function

soil_color_collection()

Method to create a general dataset collection for soil color class to use with soil albedo. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

@article{lawrence2007representing,
    author = {Lawrence, Peter J and Chase, Thomas N},
    year = {2007},
    title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
    journal = {Journal of Geophysical Research: Biogeosciences},
    volume = {112},
    pages = {G01023}
}

GriddingMachine.Collections.soil_hydraulics_collection — Function

soil_hydraulics_collection()

Method to create a general dataset collection for soil hydraulic parameters (residual soil water content - SWCR, saturated soil water content - SWCS, van Genuchten α - VGA, van Genuchten n - VGN). Supported datasets are (click to view bibtex items)

SWCR_120X_1Y_V1 (Dai et al., 2019)

SWCR_12X_1Y_V1 (regridded; Dai et al., 2019)

SWCS_120X_1Y_V1 (Dai et al., 2019)

SWCS_12X_1Y_V1 (regridded; Dai et al., 2019)

VGA_120X_1Y_V1 (Dai et al., 2019)

VGA_12X_1Y_V1 (regridded; Dai et al., 2019)

VGN_120X_1Y_V1 (Dai et al., 2019)

VGN_12X_1Y_V1 (regridded; Dai et al., 2019)

@article{dai2019global,
    author = {Dai, Yongjiu and Xin, Qinchuan and Wei, Nan and Zhang, Yonggen and Shangguan, Wei and Yuan, Hua and Zhang, Shupeng and Liu, Shaofeng and Lu, Xingjie},
	year = {2019},
    title = {A global high-resolution data set of soil hydraulic and thermal properties for land surface modeling},
	journal = {Journal of Advances in Modeling Earth Systems},
	volume = {11},
	number = {9},
	pages = {2996--3023}
}

GriddingMachine.Collections.surface_area_collection — Function

surface_area_collection()

Method to create a general dataset collection for earth surface area. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

2X_1Y_V1 (regridded; Lawrence and Chase, 2007)

@article{lawrence2007representing,
author = {Lawrence, Peter J and Chase, Thomas N},
year = {2007},
title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
journal = {Journal of Geophysical Research: Biogeosciences},
volume = {112},
pages = {G01023}
}

GriddingMachine.Collections.tree_density_collection — Function

tree_density_collection()

Method to create a general dataset collection for tree density (number of trees per area). Supported datasets are (click to view bibtex items)

120X_1Y_V1 (Crowther et al., 2017)

2X_1Y_V1 (regridded; Crowther et al., 2020)

@article{crowther2015mapping,
    author = {Crowther, Thomas W and Glick, Henry B and Covey, Kristofer R and Bettigole, Charlie and Maynard, Daniel S and Thomas, Stephen M and Smith, Jeffrey R and Hintler, Gregor and
              Duguid, Marlyse C and Amatulli, Giuseppe and others},
    year = {2015},
    title = {Mapping tree density at a global scale},
    journal = {Nature},
    volume = {525},
    number = {7568},
    pages = {201--205}
}

GriddingMachine.Collections.vcmax_collection — Function

vcmax_collection()

Method to create a general dataset collection for Vcmax. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Smith et al., 2019)

2X_1Y_V2 (Luo et al., 2019)

@article{smith2019global,
    author = {Smith, Nicholas G. and Keenan, Trevor F. and Prentice, I. Colin and Wang, Han and Wright, Ian J. and Niinemets, Ülo and Crous, Kristine Y. and Domingues, Tomas F. and
              Guerrieri, Rossella and {Yoko Ishida}, F. and Zhou, Shuangxi},
    year = {2019},
    title = {Global photosynthetic capacity is optimized to the environment},
    journal = {Ecology Letters},
    volume = {22},
    number = {3},
    pages = {506–517}
}
@article{luo2021global,
	author = {Luo, Xiangzhong and Keenan, Trevor F. and Chen, Jing M. and Croft, Holly and {Colin Prentice}, I. and Smith, Nicholas G. and Walker, Anthony P. and Wang, Han and Wang, Rong and
              Xu, Chonggang and Zhang, Yao},
	year = {2021},
	title = {Global variation in the fraction of leaf nitrogen allocated to photosynthesis},
	journal = {Nature Communications},
	volume = {12},
	number = {1},
	pages = {4866}
}

GriddingMachine.Collections.wood_density_collection — Function

wood_density_collection()

Method to create a general dataset collection for wood density. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Boonman et al., 2020)

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

Query gridded datasets

GriddingMachine.Collections.query_collection — Function

This function queries data path for a dataset, supported methods are

query_collection(ds)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:631.

query_collection(ds, version)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:647.

query_collection(artname)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:664.

GriddingMachine.Collections.query_collection — Method

query_collection(artname::String)

This method queries the local data path from given artifact name

artname Artifact name

GriddingMachine.Collections.query_collection — Method

query_collection(ds::GriddedCollection, version::String)

This method queries the local data path from collection, given

ds GriddedCollection type collection

version Queried dataset version (must be in ds.SUPPORTED_COMBOS)

Examples

dat_file = query_collection(canopy_height_collection(), "20X_1Y_V1");

GriddingMachine.Collections.query_collection — Method

query_collection(ds::GriddedCollection)

This method queries the local data path from collection for the default data, given

ds GriddedCollection type collection

Examples

dat_file = query_collection(canopy_height_collection());

Clean up collections

GriddingMachine.Collections.clean_collections! — Function

This function cleans up the collections, supported methods are

clean_collections!()
clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:697.

clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:734.

clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collections.jl:760.

GriddingMachine.Collections.clean_collections! — Method

clean_collections!(selection::String="old")

This method cleans up all selected artifacts of GriddingMachine.jl (through identify the GRIDDINGMACHINE file in the artifacts), given

selection A string indicating which artifacts to clean up

old Artifacts from an old version of GriddingMachine.jl (default)

all All Artifacts from GriddingMachine.jl

Examples

clean_collections!();
clean_collections!("old");
clean_collections!("all");

GriddingMachine.Collections.clean_collections! — Method

clean_collections!(selection::Vector{String})

This method cleans up all selected artifacts in GriddingMachine.jl, given

selection A vector of artifact names

Examples

clean_collections!(["PFT_2X_1Y_V1"]);

GriddingMachine.Collections.clean_collections! — Method

clean_collections!(selection::GriddedCollection)

This method cleans up all selected artifacts in GriddingMachine.jl, given

selection A GriddedCollection type collection

Examples

clean_collections!(pft_collection());

Indexer

GriddingMachine.Indexer.lat_ind — Function

lat_ind(lat::Number; res::Number=1)

Round the latitude and return the index in a matrix, given

lat Latitude

res Resolution in latitude

Examples

ilat = lat_ind(0.3);
ilat = lat_ind(0.3; res=0.5);

GriddingMachine.Indexer.lon_ind — Function

lon_ind(lon::Number; res::Number=1)

Round the longitude and return the index in a matrix, given

lon Longitude

res Resolution in longitude

Examples

ilon = lon_ind(90.3);
ilon = lon_ind(90.3; res=0.5);

GriddingMachine.Indexer.read_LUT — Function

This function reads look-up-table (LUT) for whole dataset or selected pieces. Supported methods are

read_LUT(fn)
read_LUT(fn, FT)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:78.

read_LUT(fn, cyc)
read_LUT(fn, cyc, FT)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:100.

read_LUT(fn, lat, lon, res)
read_LUT(fn, lat, lon, res, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:127.

read_LUT(fn, lat, lon)
read_LUT(fn, lat, lon, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:220.

read_LUT(fn, lat, lon, cyc, res)
read_LUT(fn, lat, lon, cyc, res, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:250.

read_LUT(fn, lat, lon, cyc)
read_LUT(fn, lat, lon, cyc, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:344.

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, FT::DataType = Float32)

Read the entire look-up-table from collection, given

fn Path to the target file

FT Float number type, default is Float32

Examples

read_LUT(query_collection(vcmax_collection()));
read_LUT(query_collection(vcmax_collection()), Float64);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, cyc::Int, FT::DataType = Float32)

Read the entire look-up-table from collection, given

fn Path to the target file

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

FT Float number type, default is Float32

Examples

read_LUT(query_collection(gpp_collection()), 8);
read_LUT(query_collection(gpp_collection()), 8, Float64);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116);
read_LUT(query_collection(vcmax_collection()), 30, 116, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, cyc::Int, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5, Float64);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64; interpolation=true);

read_LUT(fn::String, lat::Number, lon::Number, cyc::Int, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(gpp_collection()), 30, 116, 8);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, Float64);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, Float64; interpolation=true);

Requestor

GriddingMachine.Requestor.request_LUT — Function

This function requests data from our server for supported datasets (only latest datasets supported). Supported methods are

request_LUT(artname, lat, lon)
request_LUT(artname, lat, lon, cyc; user, interpolation, server, port)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Requestor.jl:42.

GriddingMachine.Requestor.request_LUT — Method

request_LUT(artname::String, lat::Number, lon::Number, cyc::Int = 0; user::String="Anonymous", interpolation::Bool = false, server::String = "tofu.gps.caltech.edu", port::Int = 5055)

Request data from the server, given

artname Artifact full name such as LAI_MODIS_2X_8D_2017_V1

lat Latitude

lon Longitude

cyc Cycle index, default is 0 (read entire time series)

user User name (non-registered users need to wait for 5 seconds before the server processes the request)

interpolation If true, interpolate the data linearly

server Server address

port Port number for the GriddingMachine server

Examples

request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5; interpolation=true);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5, 8);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5, 8; interpolation=true);