API

Collector

Gridded datasets

GriddingMachine.Collector.GriddedCollection — Type

struct GriddedCollection

Structure for general gridded dataset collection.

Fields

LABEL::String

Artifact label name

SUPPORTED_COMBOS::Vector{String}

Supported combinations

DEFAULT_COMBO::String

Default combination

Examples

vcmax_collection = GriddedCollection("VCMAX", ["2X_1Y_V1", "2X_1Y_V2"], "2X_1Y_V2");

GriddingMachine.Collector.biomass_collection — Function

biomass_collection()

Method to create a general dataset collection for biomass. Supported datasets are (click to view bibtex items)

ROOT_120X_1Y_V1 (Huang et al., 2021)

SHOOT_120X_1Y_V2 (Santoro et al., 2021)

@article{huang2021global,
	author = {Huang, Y. and Ciais, P. and Santoro, M. and Makowski, D. and Chave, J. and Schepaschenko, D. and Abramoff, R. Z. and Goll, D. S. and Yang, H. and Chen, Y. and Wei, W. and Piao, S.},
	year = {2021},
	title = {A global map of root biomass across the world's forests},
	journal = {Earth System Science Data},
	volume = {13},
	number = {9},
	pages = {4263–4274}
}
@article{santoro2021global,
	author = {Santoro, M. and Cartus, O. and Carvalhais, N. and Rozendaal, D. M. A. and Avitabile, V. and Araza, A. and de Bruin, S. and Herold, M. and Quegan, S. and Rodr{\'\i}guez-Veiga, P. and Balzter, H. and Carreiras, J. and Schepaschenko, D. and Korets, M. and Shimada, M. and Itoh, T. and {Moreno Mart{\'\i}nez}, {\'A}. and Cavlovic, J. and {Cazzolla Gatti}, R. and da Concei{\c c}\~ao Bispo, P. and Dewnath, N. and Labri{\`e}re, N. and Liang, J. and Lindsell, J. and Mitchard, E. T. A. and Morel, A. and {Pacheco Pascagaza}, A. M. and Ryan, C. M. and Slik, F. and {Vaglio Laurin}, G. and Verbeeck, H. and Wijaya, A. and Willcock, S.},
	year = {2021},
	title = {The global forest above-ground biomass pool for 2010 estimated from high-resolution satellite observations},
	journal = {Earth System Science Data},
	volume = {13},
	number = {8},
	pages = {3927--3950}
}

GriddingMachine.Collector.canopy_height_collection — Function

canopy_height_collection()

Method to create a general dataset collection for canopy height. Supported datasets are (click to view bibtex items)

20X_1Y_V1 (Simard et al., 2011)

2X_1Y_V2 (Boonman et al., 2020)

@article{simard2011mapping,
    author = {Simard, Marc and Pinto, Naiara and Fisher, Joshua B and Baccini, Alessandro},
    year = {2011},
    title = {Mapping forest canopy height globally with spaceborne lidar},
    journal = {Journal of Geophysical Research: Biogeosciences},
    volume = {116},
    number = {G4021}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collector.clumping_index_collection — Function

clumping_index_collection()

Method to create a general dataset collection for clumping index. Supported datasets are (click to view bibtex items)

240X_1Y_V1 (He et al., 2012)

2X_1Y_V1 (regridded; He et al., 2012)

2X_1Y_V2 (Braghiere et al., 2019)

V2 dataset are classified for different plant functional types. The indices are Broadleaf, Needleleaf, C3 grasses, C4 grasses, and shrubland.

@article{he2012global,
    author={He, Liming and Chen, Jing M and Pisek, Jan and Schaaf, Crystal B and Strahler, Alan H},
    year={2012},
    title={Global clumping index map derived from the MODIS BRDF product},
    journal={Remote Sensing of Environment},
    volume={119},
    pages={118--130}
}
@article{braghiere2019underestimation,
    author = {Braghiere, R{\'e}nato Kerches and Quaife, T and Black, E and He, L and Chen, JM},
    year = {2019},
    title = {Underestimation of global photosynthesis in Earth System Models due to representation of vegetation structure},
    journal = {Global Biogeochemical Cycles},
    volume = {33},
    number = {11},
    pages = {1358--1369}
}

GriddingMachine.Collector.elevation_collection — Function

elevation_collection()

Method to create a general dataset collection for surface elevation. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Yamazaki et al., 2017)

@article{yamazaki2017high,
    author = {Yamazaki, Dai and Ikeshima, Daiki and Tawatari, Ryunosuke and Yamaguchi, Tomohiro and O'Loughlin, Fiachra and Neal, Jeffery C and Sampson, Christopher C and Kanae, Shinjiro and
              Bates, Paul D},
    year = {2017},
    title = {A high-accuracy map of global terrain elevations},
    journal = {Geophysical Research Letters},
    volume = {44},
    number = {11},
    pages = {5844--5853}
}

GriddingMachine.Collector.gpp_collection — Function

gpp_collection()

Method to create a general dataset collection for gross primary productivity. Supported datasets are (click to view bibtex items)

MPI_RS_2X_1M_YYYY_V1 (YYYY from 2001 to 2019; Tramontana et al., 2016)

MPI_RS_2X_8D_YYYY_V1 (YYYY from 2001 to 2019; Tramontana et al., 2016)

VPM_5X_8D_YYYY_V2 (YYYY from 2000 to 2019; Zhang et al., 2017)

VPM_12X_8D_YYYY_V2 (YYYY from 2000 to 2019; Zhang et al., 2017)

@article{tramontana2016predicting,
    author = {Tramontana, Gianluca and Jung, Martin and Schwalm, Christopher R and Ichii, Kazuhito and Camps-Valls, Gustau and R{\'a}duly, Botond and Reichstein, Markus and Arain, M Altaf and
              Cescatti, Alessandro and Kiely, Gerard and others},
    year = {2016},
    title = {Predicting carbon dioxide and energy fluxes across global FLUXNET sites with regression algorithms},
    journal = {Biogeosciences},
    volume = {13},
    number = {14},
    pages = {4291--4313}
}
@article{zhang2017global,
    author = {Zhang, Yao and Xiao, Xiangming and Wu, Xiaocui and Zhou, Sha and Zhang, Geli and Qin, Yuanwei and Dong, Jinwei},
    year = {2017},
    title = {A global moderate resolution dataset of gross primary production of vegetation for 2000--2016},
    journal = {Scientific data},
    volume = {4},
    pages = {170165}
}

GriddingMachine.Collector.lai_collection — Function

lai_collection()

Method to create a general dataset collection for leaf area index. Supported datasets are (click to view bibtex items)

MODIS_2X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_2X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_10X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_10X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_20X_1M_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

MODIS_20X_8D_YYYY_V1 (YYYY from 2000 to 2020; Yuan et al., 2011)

@article{yuan2011reprocessing,
	author = {Yuan, Hua and Dai, Yongjiu and Xiao, Zhiqiang and Ji, Duoying and Shangguan, Wei},
	year = {2011},
	title = {Reprocessing the MODIS Leaf Area Index products for land surface and climate modelling},
	journal = {Remote Sensing of Environment},
	volume = {115},
	number = {5},
	pages = {1171--1187}
}

GriddingMachine.Collector.land_mask_collection — Function

land_mask_collection()

Method to create a general dataset collection for land mask. Supported datasets are (click to view bibtex items)

4X_1Y_V1 [(ERA5)]

GriddingMachine.Collector.leaf_chlorophyll_collection — Function

leaf_chlorophyll_collection()

Method to create a general dataset collection for leaf chlorophyll content. Supported datasets are (click to view bibtex items)

2X_7D_V1 (Croft et al., 2017)

@article{croft2020global,
    author = {Croft, H and Chen, JM and Wang, R and Mo, G and Luo, S and Luo, X and He, L and Gonsamo, A and Arabian, J and Zhang, Y and others},
    year = {2020},
    title = {The global distribution of leaf chlorophyll content},
    journal = {Remote Sensing of Environment},
    volume = {236},
    pages = {111479},
}

GriddingMachine.Collector.leaf_nitrogen_collection — Function

leaf_nitrogen_collection()

Method to create a general dataset collection for leaf nitrogen content. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

2X_1Y_V2 (Boonman et al., 2020)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collector.leaf_phosphorus_collection — Function

leaf_phosphorus_collection()

Method to create a general dataset collection for leaf phosphorus content. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}

GriddingMachine.Collector.pft_collection — Function

pft_collection()

Method to create a general dataset collection for plant function type ratio. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

@article{lawrence2007representing,
	author = {Lawrence, Peter J and Chase, Thomas N},
	year = {2007},
	title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
	journal = {Journal of Geophysical Research: Biogeosciences},
	volume = {112},
	pages = {G01023}
}

GriddingMachine.Collector.sif_collection — Function

sif_collection()

Method to create a general dataset collection for solar-induced chlorophyll fluorescence. Supported datasets are (click to view bibtex items)

TROPOMI_683_5X_1M_YYYY_V2 (YYYY from 2019 to 2019; Köhler et al., 2020)

TROPOMI_683_5X_8D_YYYY_V2 (YYYY from 2019 to 2019; Köhler et al., 2020)

TROPOMI_683DC_5X_1M_YYYY_V2 (YYYY from 2019 to 2019; Köhler et al., 2020)

TROPOMI_683DC_5X_8D_YYYY_V2 (YYYY from 2019 to 2019; Köhler et al., 2020)

TROPOMI_740_1X_1M_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740_1X_8D_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740_5X_1M_YYYY_V1 (YYYY from 2019 to 2020; Köhler et al., 2018)

TROPOMI_740_5X_8D_YYYY_V1 (YYYY from 2019 to 2020; Köhler et al., 2018)

TROPOMI_740_12X_1M_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740_12X_8D_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740DC_1X_1M_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740DC_1X_8D_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740DC_5X_1M_YYYY_V1 (YYYY from 2019 to 2020; Köhler et al., 2018)

TROPOMI_740DC_5X_8D_YYYY_V1 (YYYY from 2019 to 2020; Köhler et al., 2018)

TROPOMI_740DC_12X_1M_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

TROPOMI_740DC_12X_8D_YYYY_V1 (YYYY from 2018 to 2020; Köhler et al., 2018)

OCO2_757_5X_1M_YYYY_V3 (YYYY from 2014 to 2020; Sun et al., 2017)

OCO2_757DC_5X_1M_YYYY_V3 (YYYY from 2014 to 2020; Sun et al., 2017)

OCO2_771_5X_1M_YYYY_V3 (YYYY from 2014 to 2020; Sun et al., 2017)

OCO2_771DC_5X_1M_YYYY_V3 (YYYY from 2014 to 2020; Sun et al., 2017)

@article{kohler2018global,
    author = {K{\"o}hler, Philipp and Frankenberg, Christian and Magney, Troy S and Guanter, Luis and Joiner, Joanna and Landgraf, Jochen},
    year = {2018},
    title = {Global retrievals of solar-induced chlorophyll fluorescence with {TROPOMI}: {F}irst results and intersensor comparison to {OCO-2}},
    journal = {Geophysical Research Letters},
    volume = {45},
    number = {19},
    pages = {10,456--10,463}
}
@article{kohler2020global,
	author = {K{\"o}hler, Philipp and Behrenfeld, Michael J and Landgraf, Jochen and Joiner, Joanna and Magney, Troy S and Frankenberg, Christian},
	year = {2020},
	title = {Global retrievals of solar-induced chlorophyll fluorescence at red wavelengths with {TROPOMI}},
	journal = {Geophysical Research Letters},
	volume = {47},
	number = {15},
	pages = {e2020GL087541}
}
@article{sun2017oco,
	author = {Sun, Ying and Frankenberg, Christian and Wood, Jeffery D and Schimel, DS and Jung, Martin and Guanter, Luis and Drewry, DT and Verma, Manish and Porcar-Castell, Albert and Griffis, Timothy J and others},
	year = {2017},
	title = {OCO-2 advances photosynthesis observation from space via solar-induced chlorophyll fluorescence},
	journal = {Science},
	volume = {358},
	number = {6360}
}

GriddingMachine.Collector.sil_collection — Function

sil_collection()

Method to create a general dataset collection for solar-induced luminescence. Supported datasets are (click to view bibtex items)

SIL_20X_1Y_V1 (Köhler et al., 2021)

@article{kohler2021mineral,
	author = {K{\"o}hler, Philipp and Fischer, Woodward W and Rossman, George R and Grotzinger, John P and Doughty, Russell and Wang, Yujie and Yin, Yi and Frankenberg, Christian},
	year = {2021},
	title = {Mineral luminescence observed from space},
	journal = {Geophysical Research Letters},
	volume = {48},
	number = {19},
	pages = {e2021GL095227}
}

GriddingMachine.Collector.sla_collection — Function

sla_collection()

Method to create a general dataset collection for SLA (specific leaf area). Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Butler et al., 2017)

2X_1Y_V2 (Boonman et al., 2020)

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee, Arindam and Atkin, Owen K and Kattge, Jens and
              Amiaud, Bernard and others},
    year = {2017},
    title = {Mapping local and global variability in plant trait distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}
@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.Collector.soil_color_collection — Function

soil_color_collection()

Method to create a general dataset collection for soil color class to use with soil albedo. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

@article{lawrence2007representing,
    author = {Lawrence, Peter J and Chase, Thomas N},
    year = {2007},
    title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
    journal = {Journal of Geophysical Research: Biogeosciences},
    volume = {112},
    pages = {G01023}
}

GriddingMachine.Collector.soil_hydraulics_collection — Function

soil_hydraulics_collection()

Method to create a general dataset collection for soil hydraulic parameters (saturated hydraulic conductance - KSAT, residual soil water content - SWCR, saturated soil water content - SWCS, van Genuchten α - VGA, van Genuchten n - VGN). Supported datasets are (click to view bibtex items)

SWCR_120X_1Y_V1 (Dai et al., 2019)

SWCR_12X_1Y_V1 (regridded; Dai et al., 2019)

SWCS_120X_1Y_V1 (Dai et al., 2019)

SWCS_12X_1Y_V1 (regridded; Dai et al., 2019)

VGA_120X_1Y_V1 (Dai et al., 2019)

VGA_12X_1Y_V1 (regridded; Dai et al., 2019)

VGN_120X_1Y_V1 (Dai et al., 2019)

VGN_12X_1Y_V1 (regridded; Dai et al., 2019)

KSAT_100X_1Y_V2 (Gupta et al., 2021)

@article{dai2019global,
    author = {Dai, Yongjiu and Xin, Qinchuan and Wei, Nan and Zhang, Yonggen and Shangguan, Wei and Yuan, Hua and Zhang, Shupeng and Liu, Shaofeng and Lu, Xingjie},
	year = {2019},
    title = {A global high-resolution data set of soil hydraulic and thermal properties for land surface modeling},
	journal = {Journal of Advances in Modeling Earth Systems},
	volume = {11},
	number = {9},
	pages = {2996--3023}
}
@article{gupta2021global,
	author = {Gupta, Surya and Lehmann, Peter and Bonetti, Sara and Papritz, Andreas and Or, Dani},
	year = {2021},
	title = {Global Prediction of Soil Saturated Hydraulic Conductivity Using Random Forest in a Covariate-Based GeoTransfer Function (CoGTF) Framework},
	journal = {Journal of Advances in Modeling Earth Systems},
	volume = {13},
	number = {4},
	pages = {e2020MS002242}
}

GriddingMachine.Collector.surface_area_collection — Function

surface_area_collection()

Method to create a general dataset collection for earth surface area. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Lawrence and Chase, 2007)

1X_1Y_V1 (regridded; Lawrence and Chase, 2007)

@article{lawrence2007representing,
author = {Lawrence, Peter J and Chase, Thomas N},
year = {2007},
title = {Representing a new MODIS consistent land surface in the Community Land Model (CLM 3.0)},
journal = {Journal of Geophysical Research: Biogeosciences},
volume = {112},
pages = {G01023}
}

GriddingMachine.Collector.tree_density_collection — Function

tree_density_collection()

Method to create a general dataset collection for tree density (number of trees per area). Supported datasets are (click to view bibtex items)

120X_1Y_V1 (Crowther et al., 2017)

2X_1Y_V1 (regridded; Crowther et al., 2020)

@article{crowther2015mapping,
    author = {Crowther, Thomas W and Glick, Henry B and Covey, Kristofer R and Bettigole, Charlie and Maynard, Daniel S and Thomas, Stephen M and Smith, Jeffrey R and Hintler, Gregor and
              Duguid, Marlyse C and Amatulli, Giuseppe and others},
    year = {2015},
    title = {Mapping tree density at a global scale},
    journal = {Nature},
    volume = {525},
    number = {7568},
    pages = {201--205}
}

GriddingMachine.Collector.vcmax_collection — Function

vcmax_collection()

Method to create a general dataset collection for Vcmax. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Smith et al., 2019)

2X_1Y_V2 (Luo et al., 2019)

@article{smith2019global,
    author = {Smith, Nicholas G. and Keenan, Trevor F. and Prentice, I. Colin and Wang, Han and Wright, Ian J. and Niinemets, Ülo and Crous, Kristine Y. and Domingues, Tomas F. and
              Guerrieri, Rossella and {Yoko Ishida}, F. and Zhou, Shuangxi},
    year = {2019},
    title = {Global photosynthetic capacity is optimized to the environment},
    journal = {Ecology Letters},
    volume = {22},
    number = {3},
    pages = {506–517}
}
@article{luo2021global,
	author = {Luo, Xiangzhong and Keenan, Trevor F. and Chen, Jing M. and Croft, Holly and {Colin Prentice}, I. and Smith, Nicholas G. and Walker, Anthony P. and Wang, Han and Wang, Rong and
              Xu, Chonggang and Zhang, Yao},
	year = {2021},
	title = {Global variation in the fraction of leaf nitrogen allocated to photosynthesis},
	journal = {Nature Communications},
	volume = {12},
	number = {1},
	pages = {4866}
}

GriddingMachine.Collector.wood_density_collection — Function

wood_density_collection()

Method to create a general dataset collection for wood density. Supported datasets are (click to view bibtex items)

2X_1Y_V1 (Boonman et al., 2020)

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper, Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno EL and Cornelissen, Johannes HC and
              Gonzalez-Melo, Andres and Hattingh, Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

Query gridded datasets

GriddingMachine.Collector.query_collection — Function

This function queries data path for a dataset, supported methods are

query_collection(ds)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:782.

query_collection(ds, version)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:798.

query_collection(artname)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:815.

GriddingMachine.Collector.query_collection — Method

query_collection(artname::String)

This method queries the local data path from given artifact name

artname Artifact name

GriddingMachine.Collector.query_collection — Method

query_collection(ds::GriddedCollection, version::String)

This method queries the local data path from collection, given

ds GriddedCollection type collection

version Queried dataset version (must be in ds.SUPPORTED_COMBOS)

Examples

dat_file = query_collection(canopy_height_collection(), "20X_1Y_V1");

GriddingMachine.Collector.query_collection — Method

query_collection(ds::GriddedCollection)

This method queries the local data path from collection for the default data, given

ds GriddedCollection type collection

Examples

dat_file = query_collection(canopy_height_collection());

Clean up collections

GriddingMachine.Collector.clean_collections! — Function

This function cleans up the collections, supported methods are

clean_collections!()
clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:848.

clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:885.

clean_collections!(selection)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:911.

GriddingMachine.Collector.clean_collections! — Method

clean_collections!(selection::String="old")

This method cleans up all selected artifacts of GriddingMachine.jl (through identify the GRIDDINGMACHINE file in the artifacts), given

selection A string indicating which artifacts to clean up

old Artifacts from an old version of GriddingMachine.jl (default)

all All Artifacts from GriddingMachine.jl

Examples

clean_collections!();
clean_collections!("old");
clean_collections!("all");

GriddingMachine.Collector.clean_collections! — Method

clean_collections!(selection::Vector{String})

This method cleans up all selected artifacts in GriddingMachine.jl, given

selection A vector of artifact names

Examples

clean_collections!(["PFT_2X_1Y_V1"]);

GriddingMachine.Collector.clean_collections! — Method

clean_collections!(selection::GriddedCollection)

This method cleans up all selected artifacts in GriddingMachine.jl, given

selection A GriddedCollection type collection

Examples

clean_collections!(pft_collection());

Sync collections

GriddingMachine.Collector.sync_collections! — Function

This function sync the collections (only suggested to use with GriddingMachine server), supported methods are

sync_collections!(gc)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:932.

sync_collections!()

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Collector.jl:946.

GriddingMachine.Collector.sync_collections! — Method

sync_collections!(gcs::GriddedCollection)

Sync collection datasets to local drive, given

gc GriddedCollection type collection

GriddingMachine.Collector.sync_collections! — Method

sync_collections!()

Sync all datasets to local drive. This function is meant to initialize GriddingMachine server.

Indexer

GriddingMachine.Indexer.lat_ind — Function

lat_ind(lat::Number; res::Number=1)

Round the latitude and return the index in a matrix, given

lat Latitude

res Resolution in latitude

Examples

ilat = lat_ind(0.3);
ilat = lat_ind(0.3; res=0.5);

GriddingMachine.Indexer.lon_ind — Function

lon_ind(lon::Number; res::Number=1)

Round the longitude and return the index in a matrix, given

lon Longitude

res Resolution in longitude

Examples

ilon = lon_ind(90.3);
ilon = lon_ind(90.3; res=0.5);

GriddingMachine.Indexer.read_LUT — Function

This function reads look-up-table (LUT) for whole dataset or selected pieces. Supported methods are

read_LUT(fn)
read_LUT(fn, FT)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:78.

read_LUT(fn, cyc)
read_LUT(fn, cyc, FT)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:100.

read_LUT(fn, lat, lon, res)
read_LUT(fn, lat, lon, res, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:127.

read_LUT(fn, lat, lon)
read_LUT(fn, lat, lon, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:220.

read_LUT(fn, lat, lon, cyc, res)
read_LUT(fn, lat, lon, cyc, res, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:250.

read_LUT(fn, lat, lon, cyc)
read_LUT(fn, lat, lon, cyc, FT; interpolation)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Indexer.jl:344.

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, FT::DataType = Float32)

Read the entire look-up-table from collection, given

fn Path to the target file

FT Float number type, default is Float32

Examples

read_LUT(query_collection(vcmax_collection()));
read_LUT(query_collection(vcmax_collection()), Float64);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, cyc::Int, FT::DataType = Float32)

Read the entire look-up-table from collection, given

fn Path to the target file

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

FT Float number type, default is Float32

Examples

read_LUT(query_collection(gpp_collection()), 8);
read_LUT(query_collection(gpp_collection()), 8, Float64);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116);
read_LUT(query_collection(vcmax_collection()), 30, 116, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, cyc::Int, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5, Float64);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, 0.5, Float64; interpolation=true);

GriddingMachine.Indexer.read_LUT — Method

read_LUT(fn::String, lat::Number, lon::Number, res::Number, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

res Spatial resolution in °

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64);
read_LUT(query_collection(vcmax_collection()), 30, 116, 0.5, Float64; interpolation=true);

read_LUT(fn::String, lat::Number, lon::Number, cyc::Int, FT::DataType = Float32; interpolation::Bool = false)

Read the selected part of a look-up-table from collection, given

fn Path to the target file

lat Latitude in °

lon Longitude in °

cyc Cycle number, such as 8 for data in Augest in 1 1M resolution dataset

FT Float number type, default is Float32

interpolation If true, interpolate the dataset

Examples

read_LUT(query_collection(gpp_collection()), 30, 116, 8);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, Float64);
read_LUT(query_collection(gpp_collection()), 30, 116, 8, Float64; interpolation=true);

Requestor

GriddingMachine.Requestor.request_LUT — Function

This function requests data from our server for supported datasets (only latest datasets supported). Supported methods are

request_LUT(artname, lat, lon)
request_LUT(artname, lat, lon, cyc; user, interpolation, server, port)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Requestor.jl:42.

GriddingMachine.Requestor.request_LUT — Method

request_LUT(artname::String, lat::Number, lon::Number, cyc::Int = 0; user::String="Anonymous", interpolation::Bool = false, server::String = "griddingmachine.myftp.org", port::Int = 5055)

Request data from the server, given

artname Artifact full name such as LAI_MODIS_2X_8D_2017_V1

lat Latitude

lon Longitude

cyc Cycle index, default is 0 (read entire time series)

user User name (non-registered users need to wait for 5 seconds before the server processes the request)

interpolation If true, interpolate the data linearly

server Server address

port Port number for the GriddingMachine server

Examples

request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5; interpolation=true);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5, 8);
request_LUT("LAI_MODIS_2X_8D_2017_V1", 30.5, 115.5, 8; interpolation=true);

In development

Blender

GriddingMachine.Blender.truncate — Function

truncate(data::Matrix{FT}, n::Int) where {FT<:AbstractFloat}

Return a truncated a matrix, given

data Input matrix

n Integer times to truncate the matrix

Examples

mat_new = truncate(rand(10,10), 2);

GriddingMachine.Blender.regrid — Function

This function regrids the input datset to a spatially coaser one

regrid(data)
regrid(data, division)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Blender.jl:62.

regrid(data)
regrid(data, division)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Blender.jl:73.

GriddingMachine.Blender.regrid — Method

regrid(data::Matrix{FT}, division::Int = 1) where {FT<:AbstractFloat}
regrid(data::Array{FT,3}, division::Int = 1) where {FT<:AbstractFloat}

Return the regridded dataset, given

data Input dataset, 2D or 3D

division Spatial resolution is 1/division degree

Examples

data_2d = rand(720,360);
data_3d = rand(720,360,2);
new_2d = regrid(data_2d, 1);
new_3d = regrid(data_3d, 1);

Fetcher

GriddingMachine.Fetcher.ERA5LandHourly — Type

struct ERA5LandHourly

ERA5 Land hourly data from 1981 to present: cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-land

GriddingMachine.Fetcher.ERA5SingleLevelsHourly — Type

struct ERA5SingleLevelsHourly

ERA5 Single Level hourly data from 1981 to present: cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-single-levels

GriddingMachine.Fetcher.AbstractMODIS500m — Type

abstract type AbstractMODIS500m

Hierachy of AbstractMODIS500m

MOD09A1v006

MOD15A2Hv006

GriddingMachine.Fetcher.MOD15A2Hv006 — Type

struct MOD15A2Hv006 <: GriddingMachine.Fetcher.AbstractMODIS500m

MODIS/Terra Leaf Area Index/FPAR 8-Day L4 Global 500 m SIN Grid

GriddingMachine.Fetcher.MOD09A1v006 — Type

struct MOD09A1v006 <: GriddingMachine.Fetcher.AbstractMODIS500m

MODIS/Terra Surface Reflectance 8-Day L3 Global 500 m SIN Grid

GriddingMachine.Fetcher.install_cdsapi! — Function

install_cdsapi!()

Install cdsapi package if it does not exist

GriddingMachine.Fetcher.update_CDSAPI_info! — Function

update_CDSAPI_info!()

Update global user name and password for CDSAPI (note that the uid:api-key should not contain {}).

GriddingMachine.Fetcher.update_MODIS_password! — Function

update_MODIS_password!()

Update global user name and password for LP DAAC, if either of them is empty

GriddingMachine.Fetcher.fetch_data! — Function

Fetch data from the server. Supported methods are

fetch_data!(dt, year; vars, folder)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Fetcher.jl:500.

fetch_data!(dt, year; vars, folder)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Fetcher.jl:549.

fetch_data!(data_url, data_loc, year, label)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Fetcher.jl:599.

fetch_data!(dt, year)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Fetcher.jl:651.

fetch_data!(dt, year)

defined at /home/runner/work/GriddingMachine.jl/GriddingMachine.jl/src/Fetcher.jl:663.

GriddingMachine.Fetcher.fetch_data! — Method

fetch_data!(dt::ERA5LandHourly, year::Int; vars::Array{String,1} = ERA5_LAND_SELECTION, folder::String = "")

Download ERA5LandHourly dataset per year per item, given

dt ERA5LandHourly type structure

year Which year of data to download

vars Variables to download, default is ERA5_LAND_ALL

folder Folder to save the downloaded dataset, default is "" (current dir)

GriddingMachine.Fetcher.fetch_data! — Method

fetch_data!(dt::ERA5SingleLevelsHourly, year::Int; vars::Array{String,1} = ERA5_SINGLE_LEVELS_SELECTION, folder::String = "")

Download ERA5SingleLevelsHourly dataset per year per item, given

dt ERA5SingleLevelsHourly type structure

year Which year of data to download

vars Variables to download, default is ERA5_LAND_ALL

folder Folder to save the downloaded dataset, default is "" (current dir)