API

Datasets for CliMA Look-Up-Table

All data are stored in a general structure, but they are catergorized to different data types. The general data structure is

GriddingMachine.GriddedDataset — Type

struct GriddedDataset{FT<:AbstractFloat}

A general struct to store data

Fields

data

Gridded dataset

res_lat

Latitude resolution [°]

res_lon

Longitude resolution [°]

res_time

Time resolution: D-M-Y-C: day-month-year-century

var_name

Variable name

var_attr

Variable attribute

dt

Type label

Note it here that the dt field in GriddedDataset is the data identity of the stored data. Please refer the lists below.

Dataset types

Current supported dataset formats are

GriddingMachine.AbstractFormat — Type

abstract type AbstractFormat

Hierachy of AbstractFormat

FormatTIFF

FormatNC

GriddingMachine.FormatNC — Type

struct FormatNC

GriddingMachine.FormatTIFF — Type

struct FormatTIFF

General type

GriddingMachine.AbstractDataset — Type

abstract type AbstractDataset{FT}

Leaf level dataset

GriddingMachine.LeafNitrogenBoonman — Type

struct LeafNitrogenBoonman{FT}

Struct for leaf nitrogen content Link to Dataset Source

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper,
        Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno
        EL and Cornelissen, Johannes HC and Gonzalez-Melo, Andres and Hattingh,
        Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions
        at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.LeafNitrogenButler — Type

struct LeafNitrogenButler{FT}

Struct for leaf nitrogen content Link to Dataset Source

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and
        Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee,
        Arindam and Atkin, Owen K and Kattge, Jens and Amiaud, Bernard and
        others},
    year = {2017},
    title = {Mapping local and global variability in plant trait
        distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}

GriddingMachine.LeafPhosphorus — Type

struct LeafPhosphorus{FT}

Struct for leaf phosphorus content Link to Dataset Source

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and
        Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee,
        Arindam and Atkin, Owen K and Kattge, Jens and Amiaud, Bernard and
        others},
    year = {2017},
    title = {Mapping local and global variability in plant trait
        distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}

GriddingMachine.LeafSLABoonman — Type

struct LeafSLABoonman{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area) Link to Dataset Source

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper,
        Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno
        EL and Cornelissen, Johannes HC and Gonzalez-Melo, Andres and Hattingh,
        Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions
        at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.LeafSLAButler — Type

struct LeafSLAButler{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area) Link to Dataset Source

@article{butler2017mapping,
    author = {Butler, Ethan E and Datta, Abhirup and Flores-Moreno, Habacuc and
        Chen, Ming and Wythers, Kirk R and Fazayeli, Farideh and Banerjee,
        Arindam and Atkin, Owen K and Kattge, Jens and Amiaud, Bernard and
        others},
    year = {2017},
    title = {Mapping local and global variability in plant trait
        distributions},
    journal = {Proceedings of the National Academy of Sciences},
    volume = {114},
    number = {51},
    pages = {E10937--E10946}
}

GriddingMachine.VcmaxOptimalCiCa — Type

struct VcmaxOptimalCiCa{FT}

Struct for Vcmax estimated from optimal Ci:Ca ratio Link to Dataset Source

@article{smith2019global,
    author = {Smith, Nicholas G. and Keenan, Trevor F. and Prentice, I. Colin
        and Wang, Han and Wright, Ian J. and Niinemets, Ülo and Crous, Kristine
        Y. and Domingues, Tomas F. and Guerrieri, Rossella and {Yoko Ishida},
        F. and Zhou, Shuangxi},
    year = {2019},
    title = {Global photosynthetic capacity is optimized to the environment},
    journal = {Ecology Letters},
    volume = {22},
    number = {3},
    pages = {506–517}
}

Stand level datasets

GriddingMachine.CanopyHeightBoonman — Type

struct CanopyHeightBoonman{FT}

Struct for canopy height from GLAS ICESat Link to Dataset Source

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper,
        Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno
        EL and Cornelissen, Johannes HC and Gonzalez-Melo, Andres and Hattingh,
        Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions
        at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

GriddingMachine.CanopyHeightGLAS — Type

struct CanopyHeightGLAS{FT}

Struct for canopy height from GLAS ICESat Link to Dataset Source

@article{simard2011mapping,
    author = {Simard, Marc and Pinto, Naiara and Fisher, Joshua B and Baccini,
        Alessandro},
    year = {2011},
    title = {Mapping forest canopy height globally with spaceborne lidar},
    journal = {Journal of Geophysical Research: Biogeosciences},
    volume = {116},
    number = {G4021}
}

GriddingMachine.ClumpingIndexMODIS — Type

struct ClumpingIndexMODIS{FT}

Global clumping index data from MODIS BRDF 2006 Link to Dataset Source

@article{he2012global,
    author={He, Liming and Chen, Jing M and Pisek, Jan and Schaaf, Crystal B
        and Strahler, Alan H},
    year={2012},
    title={Global clumping index map derived from the MODIS BRDF product},
    journal={Remote Sensing of Environment},
    volume={119},
    pages={118--130}
}

GriddingMachine.ClumpingIndexPFT — Type

struct ClumpingIndexPFT{FT}

Struct for canopy height from GLAS ICESat, for different plant functional types. The indices are Broadleaf, Needleleaf, C3 grasses, C4 grasses, and shrubland. Data from Global clumping index data from MODIS BRDF 2006. Link to Dataset Source

@article{braghiere2019underestimation,
    author = {Braghiere, R{\'e}nato Kerches and Quaife, T and Black, E and He,
        L and Chen, JM},
    year = {2019},
    title = {Underestimation of global photosynthesis in Earth System Models
        due to representation of vegetation structure},
    journal = {Global Biogeochemical Cycles},
    volume = {33},
    number = {11},
    pages = {1358--1369}
}

GriddingMachine.GPPMPIv006 — Type

struct GPPMPIv006{FT}

Struct for MPI GPP v006 Link to Dataset Source

@article{jung2020scaling,
    author = {Jung, Martin and Schwalm, Christopher and Migliavacca, Mirco and
        Walther, Sophia and Camps-Valls, Gustau and Koirala, Sujan and Anthoni,
        Peter and Besnard, Simon and Bodesheim, Paul and Carvalhais, Nuno and
        others},
    year = {2020},
    title = {Scaling carbon fluxes from eddy covariance sites to globe:
        {S}ynthesis and evaluation of the {FLUXCOM} approach},
    journal = {Biogeosciences},
    volume = {17},
    number = {5},
    pages = {1343--1365}
}

GriddingMachine.GPPVPMv20 — Type

struct GPPVPMv20{FT}

Struct for VPM GPP v20 Link to Dataset Source

@article{zhang2017global,
    author = {Zhang, Yao and Xiao, Xiangming and Wu, Xiaocui and Zhou, Sha and
        Zhang, Geli and Qin, Yuanwei and Dong, Jinwei},
    year = {2017},
    title = {A global moderate resolution dataset of gross primary production
        of vegetation for 2000--2016},
    journal = {Scientific data},
    volume = {4},
    pages = {170165}
}

GriddingMachine.LAIMonthlyMean — Type

struct LAIMonthlyMean{FT}

Struct for monthly mean MODIS LAI Link to Dataset Source

@article{mao2019global,
    author = {Mao, J and Yan, B},
    year = {2019},
    title = {Global monthly mean leaf area index climatology, 1981-2015},
    journal = {ORNL DAAC}
}

GriddingMachine.NDVIAvhrr — Type

struct NDVIAvhrr{FT}

Struct for AVHRR NDVI

GriddingMachine.NIRoAvhrr — Type

struct NDVIAvhrr{FT}

Struct for AVHRR NIRv with offset

GriddingMachine.NIRvAvhrr — Type

struct NDVIAvhrr{FT}

Struct for AVHRR NIRv

GriddingMachine.NPPModis — Type

struct NPPModis{FT}

Struct for Modis NPP

GriddingMachine.SIFTropomi740 — Type

struct SIFTropomi740{FT}

Struct for TROPOMI SIF @ 740 nm Link to Dataset Source

@article{kohler2018global,
    author = {K{\"o}hler, Philipp and Frankenberg, Christian and Magney, Troy
        S and Guanter, Luis and Joiner, Joanna and Landgraf, Jochen},
    year = {2018},
    title = {Global retrievals of solar-induced chlorophyll fluorescence with
        {TROPOMI}: {F}irst results and intersensor comparison to {OCO-2}},
    journal = {Geophysical Research Letters},
    volume = {45},
    number = {19},
    pages = {10,456--10,463}
}

GriddingMachine.TreeDensity — Type

struct TreeDensity{FT}

Struct for tree density (number of trees per km⁻²) Link to Dataset Source

@article{crowther2015mapping,
    author = {Crowther, Thomas W and Glick, Henry B and Covey, Kristofer R and
        Bettigole, Charlie and Maynard, Daniel S and Thomas, Stephen M and
        Smith, Jeffrey R and Hintler, Gregor and Duguid, Marlyse C and
        Amatulli, Giuseppe and others},
    year = {2015},
    title = {Mapping tree density at a global scale},
    journal = {Nature},
    volume = {525},
    number = {7568},
    pages = {201--205}
}

GriddingMachine.WoodDensity — Type

struct WoodDensity{FT}

Struct for wood density (g cm⁻³) Link to Dataset Source

@article{boonman2020assessing,
    author = {Boonman, Coline CF and Ben{\'i}tez-L{\'o}pez, Ana and Schipper,
        Aafke M and Thuiller, Wilfried and Anand, Madhur and Cerabolini, Bruno
        EL and Cornelissen, Johannes HC and Gonzalez-Melo, Andres and Hattingh,
        Wesley N and Higuchi, Pedro and others},
    year = {2020},
    title = {Assessing the reliability of predicted plant trait distributions
        at the global scale},
    journal = {Global Ecology and Biogeography},
    volume = {29},
    number = {6},
    pages = {1034--1051}
}

Surface datasets

GriddingMachine.LandMaskERA5 — Type

struct LandMaskERA5{FT}

Struct for land mask from ERA5

LOAD and READ Look-Up-Table

Griddingmachine package allows for reading data from both the artifact and local files. To read the artifacts, function query_LUT is provided:

GriddingMachine.query_LUT — Function

query_LUT(dt::AbstractDataset)
query_LUT(dt::AbstractDataset, res_g::String)
query_LUT(dt::AbstractDataset, year::Int, res_g::String, res_t::String)

Query the file location from artifacts, given

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

Return the following:

Path to artifact data file

Format of the data

Variable name of NC file or Band of Tiff file

Time resolution

Whether latitude is reversed

Variable name of gridded data

Variable attribute of gridded data

A general function is provided to load the look-up tables. Before you load the table, you need to know what data you want to read (see AbstractDataset), data from which year (required for some datasets), geophysical resolution (required for some datasets), and time resolution (resuired for some datasets). Also, if you want to load local files, you need to know the file name, dataset format, and the dataset label (e.g., variable name in *.nc files, or band number in *.tif files). Be aware that some datasets may be very big, be patient when downloading the data files. Also, you need to use computers or servers with enough memory, otherwise the program may crash when loading the data.

GriddingMachine.load_LUT — Function

load_LUT(dt::AbstractDataset{FT}) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         res_g::String,
         res_t::String) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         year::Int,
         res_g::String,
         res_t::String) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         file::String,
         format::AbstractFormat,
         label::String,
         res_t::String,
         rev_lat::Bool,
         var_name::String,
         var_attr::Dict{String,String}) where {FT<:AbstractFloat}

Load look up table and return the struct, given

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

file File name to read, useful to read local files

format Dataset format from AbstractFormat

label Variable label in dataset, e.g., var name in .nc files, band numer in .tif files

rev_lat Whether latitude is stored reversely in the dataset, e.g., 90 to -90. If true, mirror the dataset on latitudinal direction

var_name Variable name of GriddedDataset

var_attr Variable attributes of GriddedDataset

Note that the artifact for GPP is about

500 MB for 0.2 degree resolution (5^2 * 36018046)

2600 MB for 0.083 degree resolution (12^2 * 36018046)

The loaded data may contain null values filled with different default values. To remove these unexpected data, a general function is provided to mask out these values.

GriddingMachine.mask_LUT! — Function

mask_LUT!(ds::GriddedDataset, default::Number)
mask_LUT!(ds::GriddedDataset, lims::Array)

Filter out the unrealistic values from the dataset, given

ds GriddedDataset type struct

default Optional. Default value to replace NaNs

lims Lower and upper limits for the dataset

A general function is provided to read the look-up tables. Note that you need to use realistic latitude, longitude, and index (required for some datasets) to read the data properly. For example, latitude must be within the range of [-90,90], longitude must be within the range of [-180,180] or [0,360], and index must be in [1,12] if you use it as month or [0,46] if you read the 8-day products. The functions lat_ind and lon_ind help convert latitude and longitude into index, respectively.

GriddingMachine.read_LUT — Function

read_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}

Read the dataset from given

ds GriddedDataset type struct

dt Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType

GriddingMachine.lat_ind — Function

lat_ind(lat::FT; res::FT) where {FT<:AbstractFloat}

Round the latitude and return the index in a matrix, Given

lat Latitude

res Resolution in latitude

GriddingMachine.lon_ind — Function

lon_ind(lon::FT; res::FT=1) where {FT<:AbstractFloat}

Round the longitude and return the index in a matrix, Given

lon Longitude

res Resolution in longitude

An alternative view method is provided in parallel with read_LUT. The difference is that there is no memory allocation using view_LUT so that changes can be made directly to the view, whereas read_LUT creates new number or arrays and changes cannot be made to original data directly.

GriddingMachine.view_LUT — Function

view_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}
view_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::Tuple{FT,FT},
         lon::Tuple{FT,FT},
         ind::Tuple{Int,Int}) where {FT<:AbstractFloat}

View the dataset from given

ds GriddedDataset type struct

dt Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType

Regrid the Look-Up-Table

For better matching the gridded data, we also provide a function to regrid the dataset.

GriddingMachine.regrid_LUT — Function

regrid_LUT(ds::AbstractDataset{FT},
           zoom::Int;
           nan_weight::Bool=false
) where {FT<:AbstractFloat}

Regrid the data from high to low resolution and return the struct, given

ds GriddedDataset type struct

zoom Integer geophysical zoom factor (>=2)

nan_weight Optional. If true, assuming nan = 0; if false, neglecting it.

Save the Look-Up-Table

To avoid unnecessary reading or regridding the dataset, a function is provided to save the dataset to .nc file

GriddingMachine.save_LUT! — Function

save_LUT!(ds::GriddedDataset{FT},
          filename::String) where {FT<:AbstractFloat}
save_LUT!(data::Array{FT,2},
          filename::String) where {FT<:AbstractFloat}
save_LUT!(data::Array{FT,3},
          filename::String,
          var_name::String,
          var_attr::Dict{String,String}) where {FT<:AbstractFloat}

Save the dataset, given

ds GriddedDataset type struct

filename NC file name to save to

data Temporaty data

var_name Variable name in nc file

var_attr Variable attributes in nc file

Note that save_LUT!(data, filename) is designed to use with temporaty data only, be cautious when using this function.

Gridding RAW data

RAW dataset types

Currently, GriddingMachine supports the following raw dataset types:

GriddingMachine.AbstractUngriddedData — Type

abstract type AbstractUngriddedData{FT}

Hierachy of AbstractUngriddedData

AbstractMODIS500m

AbstractMODIS1km

The 500 m resolution datasets include

GriddingMachine.AbstractMODIS500m — Type

abstract type AbstractMODIS500m{FT}

Hierachy of AbstractMODIS500m

MOD15A2Hv006LAI

GriddingMachine.MOD09A1v006NIRv — Type

struct MOD09A1v006LAI

Terra Surface Reflectance

GriddingMachine.MOD15A2Hv006LAI — Type

struct MOD15A2Hv006LAI

Leaf area index

The 1 km resolution datasets include

GriddingMachine.AbstractMODIS1km — Type

abstract type AbstractMODIS1km{FT}

Hierachy of AbstractMODIS1km

Steps to grid RAW data

GriddingMachine allows to grid the data in a multiple threading manner, to realize this while avoid extensive memory quota, some CSV cache files will be created during the process (with the risk of high disk usage). To make it more user friendly, a customized function is used to dynamically regulate the number of threadings:

GriddingMachine.dynamic_workers! — Function

dynamic_workers!(nthread::Int)

Change the number of workers according to CPU_THREADS, given

nthread Number of thread required

However, because the MODIS tile information file is very large (e.g., 500 m resolution tile infomation takes up to >40 GB memory to load), the tile matricies (for both latitude and longitude) will not be loaded a priori. To load the tile matricies, you will need to call load_MODIS! function manually,

GriddingMachine.load_MODIS! — Function

load_MODIS!(dt::AbstractUngriddedData) where {FT<:AbstractFloat}

Prepare parameters (file name and etc) to work on, given

dt AbstractUngriddedData type ungridded data type

Note that, if you want to load the matricies in every thread, you will need command like @everywhere load_MODIS!(MOD15A2Hv006LAI{Float32}()). What it does is loading 500 m resolution tiles information to every thread (worker). If you pass a 1 km data set type to load_MODIS!, it will load 1 km resolution tiles information automatically, you don't need to worry about it. However, be cautious to use corresponding tile information for your project.

Once the tile information is loaded, you may query the files you want to work on using query_RAW, which returns an array of paramters to pass to different threads. Note, you may need to use fetch_RAW! to download the datasets first. Function fetch_RAW! uses parse_date and update_password to download data from the website.

GriddingMachine.fetch_RAW! — Function

fetch_RAW!(dt::MOD15A2Hv006LAI, year::Int)
fetch_RAW!(data_url::String, data_loc::String, year::Int, label::String)

Download RAW product data, given

dt AbstractUngriddedData type ungridded data type

year Which year data to download

data_url Url to the website where to download the data

data_loc Local folder to store the data

GriddingMachine.parse_date — Function

parse_date(year::Int, doy::Int)

Return a string of YYYY.MM.DD, given

year Which year

doy Which day in year

GriddingMachine.query_RAW — Function

query_RAW(dt::AbstractUngriddedData,
          year::Int,
          days::Int) where {FT<:AbstractFloat}
query_RAW(dt::MOD15A2Hv006LAI,
          year::Int) where {FT<:AbstractFloat}

Prepare parameters (file name and etc) to work on, given

dt AbstractUngriddedData type ungridded data type

year AD year using format XXXX

days How many days per cycle

Note that query_RAW(dt, year) function returns the follwings

Folder in on Fluo server that contains files to process

File name pattern, such as "MOD15A2H.AYYYYDOY.*.hdf"

File name prefix, such as "MOD15A2H.A"

Band name to process, such as "Lai_500m"

Cache folder location to store generated CSV files

Scaling and offset array [a,b] as in y=ax+b

Mask of minimum and maximum values

GriddingMachine.update_password — Function

update_password()

Update global user name and password for LP DAAC, if eithe of them is empty

Next, you should be able to grid the RAW files using grid_RAW!, which extract tile information from file name using parse_HV

GriddingMachine.parse_HV — Function

parse_HV(modis_name::String)

Return the H and V index, given

modis_name MODIS file name

Then, grid_RAW! will match the data information with tile latitude and longitude, and save the data to cache CSV files, which will stay on your hard drive until you remove them manually.

GriddingMachine.grid_RAW! — Function

grid_RAW!(param::Array)
grid_RAW!(dt::MOD15A2Hv006LAI{FT},
          params::Array{Any,1}) where {FT<:AbstractFloat}
grid_RAW!(dt::MOD15A2Hv006LAI{FT},
         params::Array{Array,1},
         nthread::Int) where {FT<:AbstractFloat}

Grid the data and save it to a csv file, given

param Parameter sets to pass to different threads

dt AbstractUngriddedData type ungridded data type

params Array of param

nthread Number of threads to run the code in parallel

The last step you need is to read the cache CSV files and compile them to nc datasets. As the file may be extremely large for very high resolution data (e.g., 500 m), files are again stored as cache nc files, and then they will be compile into one nc file.

GriddingMachine.compile_RAW! — Function

compile_RAW!(param::Array)
compile_RAW!(
            dt::MOD15A2Hv006LAI{FT},
            params::Array{Any,1},
            zooms::Int
) where {FT<:AbstractFloat}

Compile RAW data to nc files, given

param Array of parameters that pass into threads

dt AbstractUngriddedData type ungridded data type

params Array of parameters from query_RAW, not that this param is different from the param used in compile_RAW!

zooms Geometrical resolution zoom factor, zooms=x means 1/x degree

A shortcut to run the gridding process is process_RAW!

GriddingMachine.process_RAW! — Function

process_RAW!(
            dt::MOD15A2Hv006LAI{FT},
            year::Int;
            days::Int = 8,
            zooms::Int = 12,
            nthread::Int = 4
) where {FT<:AbstractFloat}

Grid the data and save it to a csv file, given

param Parameter sets to pass to different threads

dt AbstractUngriddedData type ungridded data type

params Array of param

nthread Number of threads to run the code in parallel

The disadvantage is that you need a huge space to store the cache file, whereas the advantages are

all cache files can be reused, no recalculation is required if you need to grid the data to another resolution

if you encounter any error (e.g., memory shortage), you don't need to redo the calculations for cached files

you can run the gridding and compiling in multi-thread manner