API

Datasets for Clima Look-Up-Table

All data are stored in a general structure, but they are catergorized to different data types. The general data structure is

GriddingMachine.GriddedDataset — Type

struct GriddedDataset{FT<:AbstractFloat}

A general struct to store data

Fields

data

Monthly mean LAI

res_lat

Latitude resolution [°]

res_lon

Longitude resolution [°]

res_time

Time resolution: D-M-Y-C: day-month-year-century

dt

Type label

Note it here that the dt field in GriddedDataset is the data identity of the stored data. Please refer the lists below

Dataset types

Current supported dataset formats are

GriddingMachine.AbstractFormat — Type

abstract type AbstractFormat

Hierachy of AbstractFormat

FormatTIFF

FormatNC

GriddingMachine.FormatNC — Type

struct FormatNC

GriddingMachine.FormatTIFF — Type

struct FormatTIFF

General type

GriddingMachine.AbstractDataset — Type

abstract type AbstractDataset{FT}

Hierachy of AbstractDataset

AbstractCanopyHeight

AbstractClumpingIndex

AbstractGPP

AbstractLAI

AbstractLeafMN

AbstractNPP

AbstractVcmax

Canopy height

GriddingMachine.AbstractCanopyHeight — Type

abstract type AbstractCanopyHeight{FT}

Hierachy of AbstractCanopyHeight

CanopyHeightGLAS

GriddingMachine.CanopyHeightGLAS — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat

Clumping index

GriddingMachine.AbstractClumpingIndex — Type

abstract type AbstractCanopyHeight{FT}

Hierachy of AbstractCanopyHeight

ClumpingIndexMODIS

ClumpingIndexPFT

GriddingMachine.ClumpingIndexMODIS — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat

GriddingMachine.ClumpingIndexPFT — Type

struct GPPMPIv006{FT}

Struct for canopy height from GLAS ICESat, for different plant functional types. The indices are Broadleaf, Needleleaf, C3 grasses, C4 grasses, and shrubland

Gross primary productivity

GriddingMachine.AbstractGPP — Type

abstract type AbstractGPP{FT}

Hierachy of AbstractGPP

GPPMPIv006

GPPVPMv20

GriddingMachine.GPPMPIv006 — Type

struct GPPMPIv006{FT}

Struct for MPI GPP v006

GriddingMachine.GPPVPMv20 — Type

struct GPPVPMv20{FT}

Struct for VPM GPP v20

Leaf area index

GriddingMachine.AbstractLAI — Type

abstract type AbstractLAI{FT}

Hierachy of AbstractLAI

LAIMonthlyMean

GriddingMachine.LAIMonthlyMean — Type

struct LAIMonthlyMean{FT}

Struct for monthly mean MODIS LAI

Leaf mass and nutrients

GriddingMachine.AbstractLeafMN — Type

abstract type AbstractLeafMN{FT}

Hierachy of AbstractLAI

LeafNitrogen

LeafPhosphorus

LeafSLA

GriddingMachine.LeafSLA — Type

struct LeafSLA{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area)

GriddingMachine.LeafNitrogen — Type

struct LeafNitrogen{FT}

Struct for leaf nitrogen content

GriddingMachine.LeafPhosphorus — Type

struct LeafPhosphorus{FT}

Struct for leaf specific leaf area (inverse of leaf mass per area)

Net primary productivity

GriddingMachine.AbstractNPP — Type

abstract type AbstractNPP{FT}

Hierachy of AbstractLAI

NPPModis

GriddingMachine.NPPModis — Type

struct NPPModis{FT}

Struct for Modis NPP

Vcmax

GriddingMachine.AbstractVcmax — Type

abstract type AbstractLAI{FT}

Hierachy of AbstractVcmax

VcmaxOptimalCiCa

GriddingMachine.VcmaxOptimalCiCa — Type

struct VcmaxOptimalCiCa{FT}

Struct for Vcmax estimated from optimal Ci:Ca ratio

LOAD and READ Look-Up-Table

Griddingmachine package allows for reading data from both the artifact and local files. To read the artifacts, function query_LUT is provided:

GriddingMachine.query_LUT — Function

query_LUT(dt::AbstractDataset)
query_LUT(dt::AbstractDataset, year::Int, res_g::Number, res_t::String)

Query the file location from artifacts,

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

A general function is provided to load the look-up tables. Before you load the table, you need to know what data you want to read (see AbstractDataset), data from which year (required for some datasets), geophysical resolution (required for some datasets), and time resolution (resuired for some datasets). Also, if you want to load local files, you need to know the file name, dataset format, and the dataset label (e.g., variable name in *.nc files, or band number in *.tif files). Be aware that some datasets may be very big, be patient when downloading the data files. Also, you need to use computers or servers with enough memory, otherwise the program may crash when loading the data.

GriddingMachine.load_LUT — Function

load_LUT(dt::AbstractDataset{FT}) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         year::Int,
         res_g::Number,
         res_t::String) where {FT<:AbstractFloat}
load_LUT(dt::AbstractDataset{FT},
         file::String,
         format::AbstractFormat,
         label::String,
         res_t::String,
         rev_lat::Bool) where {FT<:AbstractFloat}

Load look up table and return the struct, given

dt Dataset type, subtype of AbstractDataset

year Which year

res_g Resolution in degree

res_t Resolution in time

file File name to read, useful to read local files

format Dataset format from AbstractFormat

label Variable label in dataset, e.g., var name in .nc files, band numer in .tif files

rev_lat Whether latitude is stored reversely in the dataset, e.g., 90 to -90. If true, mirror the dataset on latitudinal direction

Note that the artifact for GPP is about

500 MB for 0.2 degree resolution (5^2 * 36018046)

2600 MB for 0.083 degree resolution (12^2 * 36018046)

The loaded data may contain null values filled with different default values. To remove these unexpected data, a general function is provided to mask out these values.

GriddingMachine.mask_LUT! — Function

mask_LUT!(ds::GriddedDataset, default::Number)
mask_LUT!(ds::GriddedDataset, lims::Array)

Filter out the unrealistic values from the dataset, given

ds GriddedDataset type struct

default Optional. Default value to replace NaNs

lims Lower and upper limits for the dataset

A general function is provided to read the look-up tables. Note that you need to use realistic latitude, longitude, and index (required for some datasets) to read the data properly. For example, latitude must be within the range of [-90,90], longitude must be within the range of [-180,180] or [0,360], and index must be in [1,12] if you use it as month or [0,46] if you read the 8-day products. The functions lat_ind and lon_ind help convert latitude and longitude into index, respectively.

GriddingMachine.read_LUT — Function

read_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT) where {FT<:AbstractFloat}
read_LUT(ds::GriddedDataset{FT},
         dt::AbstractDataset{FT},
         lat::FT,
         lon::FT,
         ind::Int) where {FT<:AbstractFloat}

Read the LAI from given

ds GriddedDataset type struct

dt Dataset type, subtype of AbstractDataset

lat Latitude

lon Longitude

ind Index of cycle in the year, e.g., 1-46, or Month number from 1 to 12 for LAIMonthlyMean DataType

GriddingMachine.lat_ind — Function

lat_ind(lat::FT; res::FT) where {FT<:AbstractFloat}

Round the latitude and return the index in a matrix, Given

lat Latitude

res Resolution in latitude

GriddingMachine.lon_ind — Function

lon_ind(lon::FT; res::FT=1) where {FT<:AbstractFloat}

Round the longitude and return the index in a matrix, Given

lon Longitude

res Resolution in longitude

Regrid the Look-Up-Table

For better matching the gridded data, we also provide a function to regrid the dataset.

GriddingMachine.regrid_LUT — Function

regrid_LUT(ds::AbstractDataset{FT},
           zoom::Int;
           nan_weight::Bool=false
) where {FT<:AbstractFloat}

Regrid the data from high to low resolution and return the struct, given

ds GriddedDataset type struct

zoom Integer geophysical zoom factor (>=2)

nan_weight Optional. If true, assuming nan = 0; if false, neglecting it.